% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rectangle.r
\name{adjust_rectangle}
\alias{adjust_rectangle}
\title{Adjust rectangle}
\usage{
adjust_rectangle(points, v = 0, h = 0)
}
\arguments{
\item{points}{A list containing topleft, topright, bottomleft, and bottomright entries.
Each entry must be a names numeric vector of length two with elements "m.cy" and "m.cy",
which correspond to the x and y coordinates of each point}

\item{v, h}{Adjustment distance in vertical and horizontal direction respectively. Must
be an integer in pixels. 0 does not change anything in that direction.}
}
\description{
Increases the distance between points by some factor.
}
