% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gradient.r
\name{calculate_gradient}
\alias{calculate_gradient}
\title{Calculate gradient}
\usage{
calculate_gradient(x)
}
\arguments{
\item{x}{Image object, must be only 2-dimensional}
}
\value{
A list with elements:
\itemize{
  \item{x}{The same image passed}
  \item{gradx}{The gradient on the X-axis}
  \item{grady}{The gradient on the Y-axis}
  \item{magnitude}{The overall gradient magnitude}
  \item{orientation}{the overal gradient orientation}
}
}
\description{
Calculates gradient for a single layer
}
\author{
Sur Herrera Paredes
}
