% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images_and_masks.r
\name{get_images_and_masks}
\alias{get_images_and_masks}
\title{Get images and masks}
\usage{
get_images_and_masks(img_dir, mask_dir)
}
\arguments{
\item{img_dir}{A directory where images are located}

\item{mask_dir}{A directory containing \strong{only} mask files}
}
\value{
A lust where each element is a list with elements img and
mask. Each of those elements is an EBImage image object.
}
\description{
Makes a list of matching images and their mask. Useful to bulk display mask
images
}
\details{
Takes two directories, one for images and another for masks, then for every
file in the mask directory (mask_dir), it looks for an \strong{identically} named file in the image
directory (img_dir), creating a list of matching images and masks.
}
