% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_square.r
\name{get_square}
\alias{get_square}
\title{Get square}
\usage{
get_square(topright = NULL, topleft = NULL, bottomright = NULL,
  bottomleft = NULL)
}
\description{
Get gets the final point in the rectangle
}
