% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_set_from_image.r
\name{get_trainig_set_from_image}
\alias{get_trainig_set_from_image}
\title{Get training set from image}
\usage{
get_trainig_set_from_image(Img, n_bgpixels = 1000)
}
\description{
Get training set from image
}
