% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.r
\name{get_values}
\alias{get_values}
\title{Get feature values}
\usage{
get_values(Img)
}
\arguments{
\item{Img}{An image object. Should be from EBImage, but any three dimensional
matrix with 3 chanels specified by the last dimension should work.}
}
\value{
A data.frame with feature values for all pixels
}
\description{
From an image it calculates 3 values per color chanel.
}
\details{
It obtrains the chanel intensity, as well as the chanel intensity gradient
and gradient orientation.
}
\author{
Sur Herrera Paredes
}
