% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull_area_from_mask.r
\name{hull_area_from_masks}
\alias{hull_area_from_masks}
\title{Hull area from mask}
\usage{
hull_area_from_masks(img)
}
\arguments{
\item{img}{A two-dimensional matrix representing an image mask. Any non-zero pixel is
considered as part of the oject that wants to be encapsulated.}
}
\value{
The area (in pixels) of the convex hull
}
\description{
From an image mask, it calcultates the hull area of the unmasked object
}
\details{
The mask must be a two-dimensional matrix where 0 values indicate masked pixels,
and non-zero values unmasked (selected) pixels. The convex hull is the smallest
convex polygon that encapsulates all the unmasked pixels, and this function returns
the area of that polygon
}
\author{
Sur Herrera Paredes
}
