\name{m1_20141020}
\alias{m1_20141020}

\title{m1_20141020}

\description{ This model was trained using three images partially annotated by hand with plants grown on clay. 

It is a multiclass model that has plant soil and other classes, and achieved 99.8\% accuracy on the test.

Uses a radial kernel.}


