\name{m1_20141021tiny}
\alias{m1_20141021tiny}

\title{m1_20141021tiny}

\description{ This model was trained using three images partially annotated by hand with plants grown on clay. Then 1000 pixels were randomly selected form each of three classes

Uses a radial kernel.

includes probabilities}


