% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_image.r
\name{predict_image}
\alias{predict_image}
\title{Predict image}
\usage{
predict_image(picture, m1, size_threshold = 0, predict_class = "plant",
  outmask = "outmask/", outoverlay = "outoverlay/")
}
\description{
Predict image
}
