% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_predictdir_9feat.r
\name{wrapper_predictdir_9feat}
\alias{wrapper_predictdir_9feat}
\title{Predict from 9 features}
\usage{
wrapper_predictdir_9feat(img_dir, overlaydir, maskdir, m1)
}
\arguments{
\item{img_dir}{Directory name where images files are}

\item{overlaydir}{directory anem where to save the overlay images}

\item{maskdir}{directory name where the mask files are saved}

\item{m1}{trained SVM model from e1071 used for prediction}
}
\value{
Data.frame showing file names processed and object size in pixels
from each file
}
\description{
Takes a directory of images, and tries to predict on all images on that directory
}
\author{
Sur Herrera Paredes
}
