% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sticker.r
\name{find_sticker}
\alias{find_sticker}
\title{Find sticker}
\usage{
find_sticker(img, mins, maxs, return.img = FALSE)
}
\description{
Finds position of a sticker based on thresholds
}
\author{
Sur Herrera Paredes
}
