% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_three_stickers.r
\name{find_three_stickers}
\alias{find_three_stickers}
\title{Find the three stickers in standard positions}
\usage{
find_three_stickers(img, toplef.min = c(0.1, 0.25, 0.25),
  topleft.max = c(0.25, 0.3, 0.4), topright.min = c(0.1, 0, 0.1),
  topright.max = c(0.2, 0.15, 0.3), bottomleft.min = c(0.2, 0.1, 0.1),
  bottomleft.max = c(0.4, 0.2, 0.2), return.img = FALSE)
}
\description{
Find the three stickers in standard positions
}
