% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_masked_pixels_training_set.r
\name{get_masked_pixels_training_set}
\alias{get_masked_pixels_training_set}
\title{Get masked pixels in training set}
\usage{
get_masked_pixels_training_set(Img)
}
\description{
Get masked pixels in training set
}
