% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_set_from_image.r
\name{get_trainig_set_from_image}
\alias{get_trainig_set_from_image}
\title{Get training set from image}
\usage{
get_trainig_set_from_image(Img, n_bgpixels = 1000)
}
\arguments{
\item{Img}{A list containing *img* and *mask* elements. The former must be an three-dimensional
matrix with the original image (each RGB channel corresponding to each of the slices defined by the
third dimension), and the latter must be a two-dimensional matrix, of the exact same size as each of
the chanels from the *img* element, where non-zero values indicate the pixels of interest.}

\item{n_bgpixels}{Number of background pixels (pixels not selected in the mask) to keep.}
}
\value{
A data.frame with all the 9 feature values for the selected
pixels, as well as for the selected number of background pixels. Can
be used as training set.
}
\description{
Calculates nine features (intensity, gradient magnitude and gradient orientation,
from each of the RGB channels), and uses a mask to return the pixels of interest,
together with a specified number of pixels from the non-selected region of the image
}
\author{
Sur Herrera Paredes
}
