% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_predictdir_9feat.r
\name{wrapper_predictdir_9feat}
\alias{wrapper_predictdir_9feat}
\title{Predict fromm 9 features}
\usage{
wrapper_predictdir_9feat(img_dir, overlaydir, maskdir, m1)
}
\description{
Predict fromm 9 features
}
