% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_train_3class_9feat.r
\name{wrapper_train_3class_9feat}
\alias{wrapper_train_3class_9feat}
\title{Train on 3 classes and 9 functions}
\usage{
wrapper_train_3class_9feat(img_dir, plant_mask_dir, soil_mask_dir,
  other_mask_dir, kernel = "radial", pixels_per_class = 1e+06,
  probability = TRUE)
}
\description{
Train on 3 classes and 9 functions
}
